Move file utils to utils.py. NFC.
This allows them to be used in `config.py` (which cannot
depend on `shared.py` to avoid circular deps).
