+++ tools/shared.py
@@ -31,6 +31,7 @@ from . import cache, tempfiles, colored_logger
 from . import diagnostics
 from . import config
 from . import filelock
+from . import utils
 from .settings import settings
 
 
@@ -340,7 +341,7 @@ def set_version_globals():
 
 def generate_sanity():
   sanity_file_content = f'{EMSCRIPTEN_VERSION}|{config.LLVM_ROOT}|{get_clang_version()}'
-  config_data = read_file(config.EM_CONFIG)
+  config_data = utils.read_file(config.EM_CONFIG)
   checksum = binascii.crc32(config_data.encode())
   sanity_file_content += '|%#x\n' % checksum
   return sanity_file_content
@@ -408,7 +409,7 @@ def check_sanity(force=False):
   sanity_file = Cache.get_path('sanity.txt')
   with Cache.lock():
     if os.path.exists(sanity_file):
-      sanity_data = read_file(sanity_file)
+      sanity_data = utils.read_file(sanity_file)
       if sanity_data != expected:
         logger.debug('old sanity: %s' % sanity_data)
         logger.debug('new sanity: %s' % expected)
@@ -789,7 +790,7 @@ def read_and_preprocess(filename, expand_macros=False):
     args += ['--expandMacros']
 
   run_js_tool(path_from_root('tools/preprocessor.js'), args, True, stdout=open(stdout, 'w'), cwd=dirname)
-  out = read_file(stdout)
+  out = utils.read_file(stdout)
 
   return out
 
@@ -798,24 +799,6 @@ def do_replace(input_, pattern, replacement):
   if pattern not in input_:
     exit_with_error('expected to find pattern in input JS: %s' % pattern)
   return input_.replace(pattern, replacement)
-
-
-def read_file(file_path):
-  """Read from a file opened in text mode"""
-  with open(file_path) as fh:
-    return fh.read()
-
-
-def read_binary(file_path):
-  """Read from a file opened in binary mode"""
-  with open(file_path, 'rb') as fh:
-    return fh.read()
-
-
-def write_file(file_path, text):
-  """Write to a file opened in text mode"""
-  with open(file_path, 'w') as fh:
-    fh.write(text)
 
 
 # ============================================================================
