+++ tools/utils.py
@@ -68,3 +68,21 @@ def which(program):
             return exe_file + suffix
 
   return None
+
+
+def read_file(file_path):
+  """Read from a file opened in text mode"""
+  with open(file_path) as fh:
+    return fh.read()
+
+
+def read_binary(file_path):
+  """Read from a file opened in binary mode"""
+  with open(file_path, 'rb') as fh:
+    return fh.read()
+
+
+def write_file(file_path, text):
+  """Write to a file opened in text mode"""
+  with open(file_path, 'w') as fh:
+    fh.write(text)
