+++ tools/webassembly.py
@@ -13,6 +13,7 @@ import os
 import sys
 
 from . import shared
+from . import utils
 from .settings import settings
 
 sys.path.append(shared.path_from_root('third_party'))
@@ -97,7 +98,7 @@ def add_emscripten_metadata(wasm_file):
     #     the EMSCRIPTEN_METADATA_MINOR
   )
 
-  orig = shared.read_binary(wasm_file)
+  orig = utils.read_binary(wasm_file)
   with open(wasm_file, 'wb') as f:
     f.write(orig[0:8]) # copy magic number and version
     # write the special section
