+++ tools/webidl_binder.py
@@ -13,7 +13,7 @@ import sys
 
 sys.path.insert(1, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from tools import shared
+from tools import shared, utils
 
 sys.path.append(shared.path_from_root('third_party'))
 sys.path.append(shared.path_from_root('third_party', 'ply'))
@@ -55,7 +55,7 @@ p = WebIDL.Parser()
 p.parse(r'''
 interface VoidPtr {
 };
-''' + shared.read_file(input_file))
+''' + utils.read_file(input_file))
 data = p.finish()
 
 interfaces = {}
